refactor: change package name exceptions to exception BJS2-65898
