+++ src/main/java/school/faang/user_service/exception/EventParticipationException.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.exceptions;
+package school.faang.user_service.exception;
 
 public class EventParticipationException extends RuntimeException {
     public EventParticipationException(String message) {
